fixes for travis
